requirements.txt
python-dotenv
psycopg[binary]
pandas

